JavaScript: Move tsconfig files into `extractor.tsconfig` package
Also make the indentation in `CompilerOptions.java` more consistent.
