--- javascript/extractor/src/com/semmle/js/dependencies/tsconfig/CompilerOptions.java
@@ -1,13 +1,0 @@
-package com.semmle.js.dependencies.tsconfig;
-
-public class CompilerOptions {
-    private String outDir;
-
-    public String getOutDir() {
-        return outDir;
-    }
-
-    public void setOutDir(String outDir) {
-        this.outDir = outDir;
-    }
-}
