+++ javascript/extractor/src/com/semmle/js/extractor/tsconfig/CompilerOptions.java
@@ -1,0 +1,13 @@
+package com.semmle.js.extractor.tsconfig;
+
+public class CompilerOptions {
+  private String outDir;
+
+  public String getOutDir() {
+    return outDir;
+  }
+
+  public void setOutDir(String outDir) {
+    this.outDir = outDir;
+  }
+}
